+++ project/Source/Component/UI/ButtonUI.cpp
@@ -48,14 +48,14 @@ void ButtonUI::Update()
 		r += Time::DeltaTimeRate() * 255.0f;
 		g += Time::DeltaTimeRate() * 255.0f;
 		b += Time::DeltaTimeRate() * 255.0f;
-		if (r >= 255) {
-			r = 0;
+		if (r >= 255.0f) {
+			r = 0.0f;
 		}
-		if (g >= 255) {
-			g = 0;
+		if (g >= 255.0f) {
+			g = 0.0f;
 		}
-		if (b >= 255) {
-			b = 0;
+		if (b >= 255.0f) {
+			b = 0.0f;
 		}
 	}
 }
@@ -70,7 +70,7 @@ void ButtonUI::Draw()
 		SetDrawBright(200, 200, 200);
 	}
 	if (gradeMode) {
-		SetDrawBright(r, g, b);
+		SetDrawBright((int)r, (int)g, (int)b);
 		SetDrawBlendMode(DX_BLENDMODE_ADD, 60);
 		float add = transform->scale.x / 8;
 		for (int i = 0; i < 8; i++) {
